+++ EventListener/Swiftmailer.php
@@ -33,9 +33,6 @@ class Swiftmailer implements EventListenerInterface
 
             return;
         }
-        $newMailerId = $updatedConfiguration->getId();
-        $oldMailerId = $existingConfiguration->getId();
-        $this->updateUvdeskYmlFile($oldMailerId, $newMailerId);
 
         $mailboxConfiguration = $this->container->get('uvdesk.mailbox')->parseMailboxConfigurations(true);
 
@@ -68,8 +65,7 @@ class Swiftmailer implements EventListenerInterface
         $mailboxConfiguration = $this->container->get('uvdesk.mailbox')->parseMailboxConfigurations();
 
         foreach ($mailboxConfiguration->getMailboxes() as $existingMailbox) {
-            if(!empty($existingMailbox->getSwiftmailerConfiguration())){
-                if ($existingMailbox->getSwiftmailerConfiguration()->getId() == $configuration->getId()) {
+                if (!empty($existingMailbox->getSwiftmailerConfiguration()) && $existingMailbox->getSwiftmailerConfiguration()->getId() == $configuration->getId()) {
                 // Disable mailbox and update configuration
                     $mailbox = new Mailbox($existingMailbox->getId());
                     $mailbox->setName($existingMailbox->getName())
@@ -80,39 +76,10 @@ class Swiftmailer implements EventListenerInterface
                     $mailboxConfiguration->removeMailbox($existingMailbox);
                     $mailboxConfiguration->addMailbox($mailbox);
                 }
-            }
         }
 
         if (true === $isUpdateRequiredFlag) {
             file_put_contents($this->container->get('uvdesk.mailbox')->getPathToConfigurationFile(), (string) $mailboxConfiguration);
-        }
-        $oldMailerId = $configuration->getId();
-        $newMailerId = null;
-        // updating uvdesk.yaml file.
-        $this->updateUvdeskYmlFile($oldMailerId, $newMailerId);
-        return;
-    }
-    
-    public function updateUvdeskYmlFile($oldMailerId, $newMailerId = null)
-    {
-        $filePath = $this->container->get('kernel')->getProjectDir() . '/config/packages/uvdesk.yaml';
-        $file_content = file_get_contents($filePath);
-        $file_content_array = Yaml::parse($file_content, 6); 
-        $result = $file_content_array['uvdesk']['support_email'];
-       
-        if($result['mailer_id'] == $oldMailerId){
-            $templatePath = $this->container->get('kernel')->getProjectDir() . '/vendor/uvdesk/core-framework/Templates/uvdesk.php';
-            
-            $malierIdValue = is_null($newMailerId) ? '~' : $newMailerId;
-
-            $file_data_array = strtr(require $templatePath, [
-                '{{ SITE_URL }}' => $file_content_array['uvdesk']['site_url'],
-                '{{ SUPPORT_EMAIL_ID }}' => $file_content_array['uvdesk']['support_email']['id'] ,
-                '{{ SUPPORT_EMAIL_NAME }}' => $file_content_array['uvdesk']['support_email']['name'],
-                '{{ SUPPORT_EMAIL_MAILER_ID }}'  => $malierIdValue,
-            ]);
-            // updating contents of uvdesk.yaml file.
-            file_put_contents($filePath, $file_data_array);
         }
         return;
     }
